chore: Fully adopt the "main realm/context" terminology (#4)
It used to be that the realm/context created with the isolate was sometimes called "global context" and sometimes "main context/realm". This goes back to the days before JsRealm, when the JsRuntime::global_context method returned a v8::Global for the only v8::Context that existed in the isolate. Nowadays that is no longer the case, and we should call it the main realm.

That change by itself would also rename some properties/variables from is_global to is_main, which is also used elsewhere to refer to the top-level JsRuntime/isolate in a process. Therefore, the names are changed to is_main_runtime and is_main_realm to disambiguate.