Add documentation options for extension!() (#232)
Added docs: comma separated list of toplevel #[doc=...] tags to be applied to the extension's resulting struct to extension! to allow users to set docblocks for it, since the struct is always made public

Added comments for generated functions, since they break linting when warning about missing function docs

Added an op2 example that uses the change (also it is difficult to find samples of op2 and extension! in use with the runtime - having access to such an example would have simplified my usage of the crate significantly


---------

Co-authored-by: Matt Mastracci <matthew@mastracci.com>
Co-authored-by: Nayeem Rahman <nayeemrmn99@gmail.com>