feat: Add a re-entrancy flag for ops (#331)
Ops now generate a re-entrancy guard in debug mode that checks to see if
they are illegally re-entrant. Ideally we should limit op re-entrancy
because it can cause issues with locks, v8 states, etc.

Tests pass in Deno locally with these changes, and the given ops marked
as re-entrant.

Note that this check is compiled out in release mode for now.