feat: Add RuntimeOptions::wait_for_inspector_disconnect_callback (#387)
This commit adds
"RuntimeOptions::wait_for_inspector_disconnect_callback"
that allows to pass a callback that will be called when the event loop
runs out of work, but there are active, non-blocking inspector sessions - that is
sessions from "Chrome DevTools". This allows to print a message notifying user
that the program has finished, but is waiting for DevTools to disconnect to
exit the process.