feat: add 'requested_module_type' to ModuleLoader::load (#439)
This commit adds additional argument to "ModuleLoader::load" method that
is "RequestedModuleType". This argument can be used by the embedder to 
decide if the loaded module should be refused (eg. trying to load a
module with type "json", but the actual MIME type is not "application/json").

In a follow up I might extract `is_dyn_import` and
`requested_module_type` into a helper struct (maybe `LoadMetadata`).

Ref https://github.com/denoland/deno_core/pull/402