feat: Infrastructure for op sanitizer in deno_core (#460)
This is the underlying infrastructure we need to get the op sanitizer
working in Rust. There are no useful public-facing APIs from this yet,
but it implements the op portion of the work.

Step 2 will be adding resources and timers to the API, then we'll be
able to implement this in Deno itself.