fix: Fix segmentation fault in tests on CPUs with PKU support (#471)
Due to V8's PKU feature, only the thread that initialized the V8
platform or those spawned from it can access the isolates; other threads
will get a segmentation fault. Since each test runs in its thread, they
crash on newer CPUs that support PKU.

This issue has two possible solutions:
1. Initialize the platform once in the main thread.
2. Turn off the PKU feature during testing.

This issue was already encountered for the `deno test` command, which
was resolved by initializing the platform in the main thread
(denoland/deno#20495).

In the case of rust tests, the same solution cannot be used, as the
default test runner does not allow running code on the main thread
before any test runs.

My proposed solution is to add a feature flag to deno_core, which, when
enabled, uses the unprotected platform. An alternative solution could be
to call `deno_core::JsRuntime::init_platform` at the beginning of each
test.

This PR will also be a step towards fixing denoland/deno#21439