feat: Add ModuleLoadResponse (#491)
Checking if having this enum is beneficial and can help optimize some of
the code.

In many situations embedders will rely on `ModuleLoader::prepare_module_load`
and assemble the whole graph themselves. Then when actually loading
modules, they will wrap the already ready value into an empty future that needs
to be boxed.