feat: Use serde_json and Snapshot*DataStore for managing snapshot sidecar data (#569)
This code will be used for the future opcall tracing work, as we need to
get a reference to a number of primordials on the Rust side and persist
those references in the snapshot.

This drastically simplifies our management of snapshot sidecar data. 

As a positive side-effect, the management of module map data becomes
trivially easy. In the future, we may decide to serialize that serde_v8
data using a more efficient protocol, but that's left as a future task.

The sidecar data is serialized in the snapshot by using a single, nested
serde_v8 struct and a collection of `v8::Data` objects that are
referenced by ID. Because we cannot use serde_v8 for `v8::Data` objects
like `v8::Module`, these are simply aggregated and stored in successive
indexes of the snapshot per-context data.

It's a shame we don't have the ability to create a `v8::FixedArray`
because that would be ideal for this.

`cargo benchmark --bench snapshot` shows that the snapshot load time
difference is within the noise margin.