refactor: Use v8::OneByteConst for internals scripts (#591)
This commit changes internal JS source code of "deno_core"
to be included in the binary. We now use "v8::OneByteConst" and
pass external references to the sources which allows V8 to take
an optimized path when deserializing the snapshot. Local tests
with flamegraphs confirm that it's now cheaper to deserialize the
snapshot.

Closes https://github.com/denoland/deno_core/issues/589