serde_v8: Replace derived AsRef, AsMut for ByteString with manual impls
The new implementations are not exactly equivalent to the old ones, as
smallvec could theoretically add more AsRef and AsMut impls in addition
to the existing `[u8]` ones, that would have been picked up by serde_v8
as well then with the previous generic implementations. However, it
seems very unlikely that any such impls would ever be added and be
useful for this crate. The generated rustdocs is also easier to read
with less generic indirection.