fix: Wake event loop when dynamic import promise resolves (#769)
Fixes a hang that could occur when a module contained a dynamic import
that doesn't resolve immediately, and there was a pending op
outstanding.

The original reproduction (ref
https://github.com/denoland/deno/issues/24098) effectively boiled down
to this:

```ts
// main.ts
Deno.serve(async function (_req: Request): Promise<Response> {
    await import("./repro.ts");
    console.log("imported module");
    return new Response("done");
});

// repro.ts
await new Promise<void>((resolve) => setTimeout(resolve, 0));
console.log("module executed");
```

Upon getting a request, it would hang when trying to import
`"repro.ts"`. What was happening here was that `Deno.serve` created an
op (`op_http_serve`) that stayed pending indefinitely, so we
[assumed](https://github.com/denoland/deno_core/blob/4beb54e862805be6b2c4d8fb2bfc49ea72705799/core/runtime/jsruntime.rs#L1816C6-L1822C9)
that we would poll again, but that didn't actually hold. Unlike
non-dynamic module evaluation, we weren't waking the event loop for
polling when the dynamic import evaluation promise resolved.

The fix here is to attaching callbacks to the dynamic import promise to
wake the event loop when it resolves, similar to what we do for normal
module evaluation.