feat: add Option<GcResource> support in ops
Optionals are supported both as arguments, via
`#[cppgc] resource: Option<&T>`, and as return
values, via `Option<T>`.

This commit also adds the `GcResource` trait,
which all CppGc resources must implement. This is
done to ensure that one does not accidentally
create CppGc objects for resources that one did
not intend to be managed by CppGc.