fix: remove future machiney from evaluate_sync (#910)
Fixes crash reported in
https://github.com/denoland/deno/issues/20610#issuecomment-2357238775

The assert was intended to mean that the promise returned from evaluate
should always be settled, but instead we were asserting that the
mpsc::channel never needed to be polled more than once. Also fixes an
issue where we were draining the microtask queue within
`op_import_sync`, which is wrong.