fix: correctly trace cppgc proto objects (#1113)
The bug specifically was this code: `ptr.ptr.trace(visitor);`. It is not
correct to directly call `foo.trace(visitor)`, since you are skipping
the visitor's marking logic when you do that. You have to do
`visitor.trace(&foo)`.

In order to fix this, the impl of `ErasedPtr` was switched to use
`Member<T>`, which is in fact the correct type to use for strong
references to other cppgc objects.

Fixes: https://github.com/denoland/deno_core/issues/1112