fix(compile): panic when calling op_is_terminal (#1116)
This works towards fixing
[#27730](https://github.com/denoland/deno/issues/27730) and
[#21091](https://github.com/denoland/deno/issues/21091).

An error is not handled when calling `op_is_terminal`.
Ironically, this causes the application to panic when bootstrapping if
there is no terminal in windows.

This PR does not fully fix the issue yet, as there is a similar problem
in deno's code base, see this [pull
request](https://github.com/denoland/deno/pull/28823).