Add "text" and "bytes" module types (#1025)
This commit adds support for "text" and "bytes" modules that can
be loaded like so:

```
import txt from "./log.txt" with {type: "text"};
import bytes from "./img.png" with {type: "bytes"};
```

The necessary infrastructure is now included in `deno_core` to
instantiate and execute these modules but it is up to the 
embedder to properly implement `ModuleLoader` to return
relevant source for each module type.

`ModuleLoader::prepare_module_load` was changed to include
`RequestedModuleType` argument that can be useful to skip
unnecessary work when loading non-JS module types.
---------

Co-authored-by: Bartek Iwańczuk <biwanczuk@gmail.com>