feat(BREAKING): Make `LocalInspectorSession` synchronous (#1193)
Required for https://github.com/denoland/deno/pull/30743

This commit changes `LocalInspectorSession` to only provide a
sync API.

While it might feel counterintuitive, as it is now the callers
responsibility to handle request/response cycles, this change actually greatly
simplifies inspector implementation and gives granular control
over behavior of inspector. The main change is that sync sessions, no
longer require event loop ticks to make progress - ie. sending a message
is fully synchronous - because there no channels involved. This leads
to the fact that "responses" (which are actually callbacks) are most
often produces when the message is sent and returned. 

Thanks to this change, it's now possible to easily implement
interactions with V8 coverage collection and profiler inspector domains.