refactor: remove rules! macro usage (#1239)
We are removing the rules! macro due to its complex output and behaviour
(which involves converting token to other token types), inaddition to
its panicking and catch unwinding needs.
In its stead we are now using syn to break down the types as we need
them, with exception of one case where we have a large match on various
different types, where we now use a new crate called
[`syn-match`](https://crates.io/crates/syn-match/), which has the whole
purpouse of matching on `syn::Path`, and doesnt convert tokens to other
types, but lets you bind sections of a `syn::Path` to its corresponding
component.